+++ src/parts/pintable.h
@@ -528,9 +528,9 @@ public:
    void GetViewRect(FRect *pfrect) const;
 
    bool IsNameUnique(const WCHAR *const wzName) const;
-   void GetUniqueName(const ItemTypeEnum type, WCHAR *const wzUniqueName, const DWORD wzUniqueName_maxlength) const;
-   void GetUniqueName(const WCHAR *const wzRoot, WCHAR *const wzUniqueName, const DWORD wzUniqueName_maxlength) const;
-   void GetUniqueNamePasting(const int type, WCHAR *const wzUniqueName, const DWORD wzUniqueName_maxlength) const;
+   void GetUniqueName(const ItemTypeEnum type, WCHAR *const wzUniqueName, const size_t wzUniqueName_maxlength) const;
+   void GetUniqueName(const WCHAR *const wzRoot, WCHAR *const wzUniqueName, const size_t wzUniqueName_maxlength) const;
+   void GetUniqueNamePasting(const int type, WCHAR *const wzUniqueName, const size_t wzUniqueName_maxlength) const;
 
    float GetSurfaceHeight(const string &name, float x, float y) const;
